Fixed error message for duplicate typeIds
I was getting the error `There is already a TypeAdapter for typeId 58` when in reality I should have been getting the error `There is already a TypeAdapter for typeId 26`. This PR simply makes sure the error developers see takes the reserved type ids into account so that it matches the typeId they actually used in their adapter.